add: medium widget design barebone
