--- ios/today classes widget/today_classes_widget.swift
@@ -1,68 +1,0 @@
-//
-//  today_classes_widget.swift
-//  today classes widget
-//
-//  Created by Soongyu Kwon on 28/03/2023.
-//  Copyright © 2023 The Chromium Authors. All rights reserved.
-//
-
-import WidgetKit
-import SwiftUI
-
-struct Provider: TimelineProvider {
-    func placeholder(in context: Context) -> SimpleEntry {
-        SimpleEntry(date: Date())
-    }
-
-    func getSnapshot(in context: Context, completion: @escaping (SimpleEntry) -> ()) {
-        let entry = SimpleEntry(date: Date())
-        completion(entry)
-    }
-
-    func getTimeline(in context: Context, completion: @escaping (Timeline<Entry>) -> ()) {
-        var entries: [SimpleEntry] = []
-
-        // Generate a timeline consisting of five entries an hour apart, starting from the current date.
-        let currentDate = Date()
-        for hourOffset in 0 ..< 5 {
-            let entryDate = Calendar.current.date(byAdding: .hour, value: hourOffset, to: currentDate)!
-            let entry = SimpleEntry(date: entryDate)
-            entries.append(entry)
-        }
-
-        let timeline = Timeline(entries: entries, policy: .atEnd)
-        completion(timeline)
-    }
-}
-
-struct SimpleEntry: TimelineEntry {
-    let date: Date
-}
-
-struct today_classes_widgetEntryView : View {
-    var entry: Provider.Entry
-
-    var body: some View {
-        Text("오늘 수업 위젯")
-    }
-}
-
-struct today_classes_widget: Widget {
-    let kind: String = "today_classes_widget"
-
-    var body: some WidgetConfiguration {
-        StaticConfiguration(kind: kind, provider: Provider()) { entry in
-            today_classes_widgetEntryView(entry: entry)
-        }
-        .configurationDisplayName("오늘 수업")
-        .description("오늘 진행할 수업을 확인합니다.")
-        .supportedFamilies([.systemMedium])
-    }
-}
-
-struct today_classes_widget_Previews: PreviewProvider {
-    static var previews: some View {
-        today_classes_widgetEntryView(entry: SimpleEntry(date: Date()))
-            .previewContext(WidgetPreviewContext(family: .systemSmall))
-    }
-}
