--- ios/today classes widget/today_classes_widgetBundle.swift
@@ -1,17 +1,0 @@
-//
-//  today_classes_widgetBundle.swift
-//  today classes widget
-//
-//  Created by Soongyu Kwon on 28/03/2023.
-//  Copyright © 2023 The Chromium Authors. All rights reserved.
-//
-
-import WidgetKit
-import SwiftUI
-
-@main
-struct today_classes_widgetBundle: WidgetBundle {
-    var body: some Widget {
-        today_classes_widget()
-    }
-}
