--- ios/week classes widget/week_classes_widget.swift
@@ -1,68 +1,0 @@
-//
-//  week_classes_widget.swift
-//  week classes widget
-//
-//  Created by Soongyu Kwon on 28/03/2023.
-//  Copyright © 2023 The Chromium Authors. All rights reserved.
-//
-
-import WidgetKit
-import SwiftUI
-
-struct Provider: TimelineProvider {
-    func placeholder(in context: Context) -> SimpleEntry {
-        SimpleEntry(date: Date())
-    }
-
-    func getSnapshot(in context: Context, completion: @escaping (SimpleEntry) -> ()) {
-        let entry = SimpleEntry(date: Date())
-        completion(entry)
-    }
-
-    func getTimeline(in context: Context, completion: @escaping (Timeline<Entry>) -> ()) {
-        var entries: [SimpleEntry] = []
-
-        // Generate a timeline consisting of five entries an hour apart, starting from the current date.
-        let currentDate = Date()
-        for hourOffset in 0 ..< 5 {
-            let entryDate = Calendar.current.date(byAdding: .hour, value: hourOffset, to: currentDate)!
-            let entry = SimpleEntry(date: entryDate)
-            entries.append(entry)
-        }
-
-        let timeline = Timeline(entries: entries, policy: .atEnd)
-        completion(timeline)
-    }
-}
-
-struct SimpleEntry: TimelineEntry {
-    let date: Date
-}
-
-struct week_classes_widgetEntryView : View {
-    var entry: Provider.Entry
-
-    var body: some View {
-        Text("시간표 위젯")
-    }
-}
-
-struct week_classes_widget: Widget {
-    let kind: String = "week_classes_widget"
-
-    var body: some WidgetConfiguration {
-        StaticConfiguration(kind: kind, provider: Provider()) { entry in
-            week_classes_widgetEntryView(entry: entry)
-        }
-        .configurationDisplayName("시간표")
-        .description("수업 시간표를 확인합니다.")
-        .supportedFamilies([.systemLarge])
-    }
-}
-
-struct week_classes_widget_Previews: PreviewProvider {
-    static var previews: some View {
-        week_classes_widgetEntryView(entry: SimpleEntry(date: Date()))
-            .previewContext(WidgetPreviewContext(family: .systemSmall))
-    }
-}
