--- ios/week classes widget/week_classes_widgetBundle.swift
@@ -1,17 +1,0 @@
-//
-//  week_classes_widgetBundle.swift
-//  week classes widget
-//
-//  Created by Soongyu Kwon on 28/03/2023.
-//  Copyright © 2023 The Chromium Authors. All rights reserved.
-//
-
-import WidgetKit
-import SwiftUI
-
-@main
-struct week_classes_widgetBundle: WidgetBundle {
-    var body: some Widget {
-        week_classes_widget()
-    }
-}
